lock

+++ cmd/lotus-bench/main.go
@@ -292,6 +292,9 @@ var sealBenchCmd = &cli.Command{
 			return xerrors.Errorf("only-p1 and only-p2 cannot be true at the same time")
 		}
 
+		if os.Getenv("FIL_PROOFS_SSD_PARENT") == "" {
+			panic("FIL_PROOFS_SSD_PARENT not set")
+		}
 		sectorstorage.ShellExecute("rm -rf " + filepath.Join(os.Getenv("FIL_PROOFS_SSD_PARENT"), "*"))
 
 		os.Mkdir(filepath.Join(sbdir, "cache"), 0755)
